Privatize dfs modules
