+++ mikeio/_spectral.py
