+++ mikeio/dataset/_data_plot.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from ..spatial.FM_utils import _plot_map, _plot_vertical_profile
 
-from ..spectral import plot_2dspectrum
+from .._spectral import plot_2dspectrum
 
 
 class _DataArrayPlotter:
