+++ mikeio/dataset/_dataset.py
@@ -976,7 +976,7 @@ class Dataset(MutableMapping):
             A dataset with data dimension t
             The first two items will be x- and y- coordinates of track
         """
-        from ..track import _extract_track
+        from .._track import _extract_track
 
         item_numbers = list(range(self.n_items))
         time_steps = list(range(self.n_timesteps))
@@ -1834,18 +1834,18 @@ class Dataset(MutableMapping):
 
     def _to_dfs2(self, filename):
         # assumes Grid2D geometry
-        from ..dfs2 import write_dfs2
+        from ..dfs._dfs2 import write_dfs2
 
         write_dfs2(filename, self)
 
     def _to_dfs3(self, filename):
         # assumes Grid3D geometry
-        from ..dfs3 import write_dfs3
+        from ..dfs._dfs3 import write_dfs3
 
         write_dfs3(filename, self)
 
     def _to_dfs1(self, filename):
-        from ..dfs1 import Dfs1
+        from ..dfs._dfs1 import Dfs1
 
         dfs = Dfs1()
         dfs.write(filename, data=self, dx=self.geometry.dx, x0=self.geometry._x0)
