+++ mikeio/dfsu/_spectral.py
@@ -7,8 +7,8 @@ from mikecore.DfsuFile import DfsuFile, DfsuFileType
 from tqdm import trange
 
 from ..dataset import DataArray, Dataset
-from ..dfs import _get_item_info, _valid_item_numbers, _valid_timesteps
-from ..spectral import calc_m0_from_spectrum, plot_2dspectrum
+from ..dfs._dfs import _get_item_info, _valid_item_numbers, _valid_timesteps
+from .._spectral import calc_m0_from_spectrum, plot_2dspectrum
 from ._dfsu import _Dfsu
