+++ tests/test_dfs2.py
@@ -9,8 +9,7 @@ import xarray
 
 import mikeio
 
-from mikeio.dataset import Dataset
-from mikeio.dfs2 import Dfs2
+from mikeio import Dataset, Dfs2
 from mikeio.eum import EUMType, ItemInfo, EUMUnit
 from mikeio.exceptions import ItemsError
 from mikeio.spatial.geometry import GeometryPoint2D
