+++ tests/test_dfsu_spectral.py
@@ -6,7 +6,7 @@ from mikeio import eum
 from mikecore.DfsuFile import DfsuFileType
 
 from mikeio.spatial.FM_geometry import GeometryFMPointSpectrum, GeometryFMAreaSpectrum
-import mikeio.spectral as spectral
+import mikeio._spectral as _spectral
 
 
 @pytest.fixture
@@ -329,7 +329,7 @@ def test_read_spectrum_dir_line(dfsu_line_dir):
 def test_calc_frequency_bin_sizes(dfsu_line):
     dfs = dfsu_line
     f = dfs.frequencies
-    df = spectral._f_to_df(f)
+    df = _spectral._f_to_df(f)
     assert len(f) == len(df)
     assert df.max() < f.max()
